make time units more flexible
git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@12850 fc592f45-0881-4044-91ff-c96495a54512

+++ extra/ncArray/inst/ncreadtime.m
@@ -7,7 +7,7 @@
 
 function t = ncreadtime(filename,varname)
 
-t = ncread(filename,varname);
+t = double(ncread(filename,varname));
 units = ncreadatt(filename,varname,'units');
 
 [t0,f] = nctimeunits(units);
